# internal/handler/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Model Router",
    "description": "OpenAI-compatible proxy that groups model backends and enforces per-consumer concurrency limits.",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/chat/completions": {
      "post": {
        "summary": "Create a chat completion",
        "description": "Proxied to the selected backend of the model named in the request body. Supports streaming via \"stream\": true (SSE).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatCompletionRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Completion response (JSON or SSE stream)." },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "429": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/completions": {
      "post": {
        "summary": "Create a completion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompletionRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Completion response (JSON or SSE stream)." },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "429": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/embeddings": {
      "post": {
        "summary": "Create embeddings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmbeddingsRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Embeddings response." },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "429": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/models": {
      "get": {
        "summary": "List available models",
        "description": "Returns only models that currently have at least one healthy backend.",
        "responses": {
          "200": { "description": "A list of available models." },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Router liveness probe",
        "responses": { "200": { "description": "Router is alive." } }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Prometheus metrics",
        "responses": { "200": { "description": "Prometheus exposition format." } }
      }
    }
  },
  "components": {
    "responses": {
      "Error": {
        "description": "Error response.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["model", "messages"],
        "properties": {
          "model": { "type": "string" },
          "messages": { "type": "array", "items": { "type": "object" } },
          "stream": { "type": "boolean" }
        }
      },
      "CompletionRequest": {
        "type": "object",
        "required": ["model", "prompt"],
        "properties": {
          "model": { "type": "string" },
          "prompt": {},
          "stream": { "type": "boolean" }
        }
      },
      "EmbeddingsRequest": {
        "type": "object",
        "required": ["model", "input"],
        "properties": {
          "model": { "type": "string" },
          "input": {}
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string" },
              "type": { "type": "string" },
              "code": { "type": "integer" }
            }
          }
        }
      }
    }
  }
}